terraform/tq_creds/requirements.txt
random-password-generator==2.2.0